+++ Kmarket/src/main/java/kr/co/kmarket/dto/ProductOrderDTO.java
@@ -1,6 +1,6 @@
 package kr.co.kmarket.dto;
 
-public class ProductOrder {
+public class ProductOrderDTO {
 	private int ordNo;
 	private String ordUid;
 	private int ordCount;
