+++ Kmarket/src/main/java/kr/co/kmarket/dto/ProductReviewDTO.java
@@ -1,6 +1,6 @@
 package kr.co.kmarket.dto;
 
-public class ProductReview {
+public class ProductReviewDTO {
 	private int revNo;
 	private int prodNo;
 	private String content;
